refactor: simplify formatter
